+++ src/lint/latex-formatter/latexindent.ts
@@ -170,7 +170,7 @@ function format(document: vscode.TextDocument, range?: vscode.Range): Thenable<v
         })
         worker.on('error', err => {
             removeTemporaryFiles()
-            void logger.showErrorMessage('Formatting failed. Please refer to LaTex Secure Workspace Output for details.')
+            void logger.showErrorMessage('Formatting failed. Please refer to LaTeX-Secure-Workspace Output for details.')
             logger.log(`Formatting failed: ${err.message}`)
             logger.log(`stderr: ${Buffer.concat(stderrBuffer).toString()}`)
             resolve(undefined)
@@ -178,7 +178,7 @@ function format(document: vscode.TextDocument, range?: vscode.Range): Thenable<v
         worker.on('close', code => {
             removeTemporaryFiles()
             if (code !== 0) {
-                void logger.showErrorMessage('Formatting failed. Please refer to LaTex Secure Workspace Output for details.')
+                void logger.showErrorMessage('Formatting failed. Please refer to LaTeX-Secure-Workspace Output for details.')
                 logger.log(`Formatting failed with exit code ${code}`)
                 logger.log(`stderr: ${Buffer.concat(stderrBuffer).toString()}`)
                 return resolve(undefined)